Part 1
Part 1 of the MADA Semester Long Project. Edited the All of my edits were made in the Manuscript.qmd file located in the manuscript folder located in the products folder.  In the manuscript.qmd file I outlined Part 1 of the project. I also edited the processingcode.R file but that will be subject to change.